Created the home view to render the home.html template with all the dog_products.
Created the dog_product_detail view to render the dog_product_detail template that displays the chosen product's details.

Created the purchase_dog_product view that takes the chosen product and takes one off fo the quantity if it is in stock.

Created the purchase_detail view that renders the purchase_detail template along with the product purchased.

Created the new_dog_tag view that takes the form NewDogTagForm and validates the information provided. If the information is valid then it redirects to the dog_tag_list otherwise it renders the new_dog_tag template with the form.

Cerated the dog_tag_list veiw that renders the dog_tag_list template and displays all of the dog tags.